args.yaml test now working

+++ tests/mhelpers.py
@@ -20,7 +20,7 @@ MOCK_PARAM_CHI_DICT = {"param_two": "two",
                        "var_param": MOCK_CHI_PARAM_STR}
 
 MOCK_NON_YAML_FILE_STR = "This is a tests."
-MOCK_ARGS_FILE_lIST = {'stage1': ['arg1', 'arg2', 'arg3'],
+MOCK_ARGS_FILE_DICT = {'stage1': ['arg1', 'arg2', 'arg3'],
                        'stage2': ['arg4', 'arg5', 'arg6']}
 
 MOCK_PARAM_DICT_PATH = 'mock_param.yaml'
@@ -61,12 +61,13 @@ def setup_and_teardown(clean_up):
 
 
 @pytest.fixture()
-def mock_args_file(mock_root_dir):
-    args_file_path = mock_root_dir / 'args_file.yaml'
+def mock_args_file(mock_root_dir, clean_up):
+    args_file_path = mock_root_dir / 'args.yaml'
     with args_file_path.open('w') as aff:
-        yaml.dump(MOCK_ARGS_FILE_lIST, aff)
+        yaml.dump(MOCK_ARGS_FILE_DICT, aff)
     yield args_file_path
-    args_file_path.unlink()
+    if clean_up:
+        args_file_path.unlink()
 
 
 @pytest.fixture()
